Fix circular dependency problem with the tests
