+++ server/api/user.ts
@@ -1,6 +1,6 @@
 import * as db from './db';
 import * as utils from './utils';
-import { badRequest, jsonResponse, internalError, successResponse } from './utils';
+import { badRequest, jsonResponse, internalError, successResponse, unauthorizedError, AuthToken } from './requestUtils';
 import * as models from './models';
 import * as express from 'express';
 import * as security from './security';
@@ -97,7 +97,7 @@ export async function handleLogin(req: express.Request, res: express.Response): 
                 },
                 left: async error => {
                     if (error === db.DatabaseErrorMessage.NOT_FOUND) {
-                        utils.unauthorizedError(res, 'could not find user');
+                        unauthorizedError(res, 'could not find user');
                     } else {
                         badRequest(res);
                     }
@@ -113,7 +113,7 @@ export async function handleLogin(req: express.Request, res: express.Response): 
     }
 }
 
-export async function handleGetAccount(req: express.Request, res: express.Response, token: security.AuthToken): Promise<void> {
+export async function handleGetAccount(req: express.Request, res: express.Response, token: AuthToken): Promise<void> {
     const userResult = await db.getUserFromEmail({email: token.email});
     userResult.caseOf({
         left: e => badRequest(res, 'user not found'),
@@ -147,7 +147,7 @@ export async function handleVerify(req: express.Request, res: express.Response):
 
 export async function handleSubscribe(req: express.Request,
                                       res: express.Response,
-                                      authToken: security.AuthToken,
+                                      authToken: AuthToken,
                                       tripType: db.AddToCampusOrAirport
                                      ): Promise<void> {
     if (!req.body || !req.body.tripDate ||
