+++ server/spec/user_spec.ts
@@ -1,6 +1,5 @@
 import * as models from '../api/models';
-import { IGetBackResponse } from '../api/utils';
-import { AuthToken } from '../api/security';
+import { IGetBackResponse, AuthToken } from '../api/requestUtils';
 import * as test_utils from './test_utils';
 import * as WebRequest from 'web-request';
